// webapp/src/api/model/eItemEquipmentManipulator.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type EItemEquipmentManipulator = typeof EItemEquipmentManipulator[keyof typeof EItemEquipmentManipulator];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EItemEquipmentManipulator = {
  SLOTS: 'SLOTS',
  DAMAGE: 'DAMAGE',
  HIT: 'HIT',
  INITIATIVE: 'INITIATIVE',
  ARMOR: 'ARMOR',
  WEIGHT: 'WEIGHT',
  PROTECTION: 'PROTECTION',
} as const;