// webapp/src/api/model/eUpgradeRestriction.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type EUpgradeRestriction = typeof EUpgradeRestriction[keyof typeof EUpgradeRestriction];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EUpgradeRestriction = {
  ITEM: 'ITEM',
  EQUIPMENT: 'EQUIPMENT',
  JEWELLERY: 'JEWELLERY',
  WEAPON: 'WEAPON',
  DEFENSIVE_ITEM: 'DEFENSIVE_ITEM',
  SHIELD: 'SHIELD',
  HANDHELD: 'HANDHELD',
  ARMOR: 'ARMOR',
} as const;